Switched to Excon as the HTTP client, added VCR
The reason for the change was the the previous version of the gem
essentially stubbed out all the HTTP requests. The service has since
then been released with a changed API. In order to test the service as
realistically as possible, added VCR where applicable.

The only interaction I'm not too confident about is the timeout. Will
see what the others say, but currently this is my best guess.